create new component searchPostControl
