+++ src/components/handpicked-inspector-controls/handpicked-inspector-controls.scss
@@ -9,7 +9,7 @@ h5.splx-selectedSearchResultTitle {
 
 .splx-selectedSearchResults .components-card {
   margin-bottom: 0.5em;
-  
+
   .components-card__body {
     padding: 0.7em;
   }
@@ -17,30 +17,4 @@ h5.splx-selectedSearchResultTitle {
 
 .splx-selectedSearchResultsWrap {
   margin-top: 3em;
-}
-
-.splx-searchResult {
-  display: flex;
-  > div {
-    em {
-      font-style: normal;
-    }
-    span {
-      display: block;
-      font-size: 80%;
-    }
-  }
-  > button {
-    margin-left: auto;
-    align-self: center;
-  }
-  > span {
-    display: inline-block;
-    margin-right: 0.5em;
-  }
-  & + & {
-    margin-top: 0.5em;
-    padding-top: 0.5em;
-    border-top: 1px solid rgb(226, 228, 231);
-  }
 }
